#108 Logo and address is now on each page.
Along with the definition of one address in letterhead by AddressWriter, the invoiceSpecific_address. from ReportDto became useless: The only address in letterhead should be also the invoice address (default address). The only deviating address in all reports is the address in delivery notes (shipping address). Consequently using invoiceSpecific_address is confusing.

GUI test resulted in no addresses in letterheads. Will have to fix that in the next commit.